+++ src/main/java/baseball/controller/BaseballGameController.java
@@ -1,5 +1,10 @@
-package baseball;
+package baseball.controller;
 
+import baseball.model.BaseballNumber;
+import baseball.model.BaseballNumbersFactory;
+import baseball.utils.Validation;
+import baseball.view.InputView;
+import baseball.view.OutputView;
 import java.util.List;
 
 public class BaseballGameController {
