+++ src/test/java/baseball/BaseballNumbersFactoryTest.java
@@ -3,6 +3,8 @@ package baseball;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
 
+import baseball.model.BaseballNumber;
+import baseball.model.BaseballNumbersFactory;
 import camp.nextstep.edu.missionutils.test.NsTest;
 import java.util.ArrayList;
 import java.util.List;
